Expose charsets as Hash in normalized params
Parse the JSON string and return a Hash with symbol keys
for ease of use by Griddler later on. This is useful for encoding
strings properly.

We are currently running into an issue where we receive text bodies
containing non-UTF8 characters. We want to convert them to UTF8
using String#encode, but without knowing the source encoding, we
cannot do this correctly. This surfaces the encodings that were
pulled out by SendGrid so that we can access them from the
Griddler Email.